Fix additional warnings given by gcc 4.8 -m32

+++ proccontrol/src/loadLibrary/codegen-ppc.C
@@ -79,7 +79,7 @@ void Codegen::generatePPC32(Address val, unsigned reg) {
 }
 
 void Codegen::generatePPC64(Address val, unsigned reg) {
-   generatePPC32(val >> 32, reg);
+   generatePPC32((uint64_t)val >> 32, reg);
 
    instruction insn;
    insn.clear();
